Control pass on Sec. 4
Several pages: consistent use of “Example” and “Examples” for one and
more examples following.
P. 291: “… any of the values in the OMP_PROC_BIND environment variable
…” -> “… the value of the OMP_PROC_BIND environment variable …”, since
there is only one value
P. 293: “… values of list accepted for the OMP_PLACES environment
variable.” -> “… values accepted …” since list was never defined
p. 293: fixed grammar, so that it is consistent with the wording and
does no longer allow OMP_PLACES=”cores,threads,{0-3}:2”, see mal thread
“[Omp-lang] Question on ticket #286” for details.
p. 296: removed linebreak in the middle of the sentence
p 300: OMP_TASK_MAX_PRIORITY -> OMP_MAX_TASK_PRIORITY was misspelled in
the text
p 300: removed “The default value for this environment variable is
zero.”, because the default values are for ICVs and described in the ICV
section. No other environment variable has a default value.