more tweaks to depend clause section
Replacing ", where" with "when" for description of task dependences and loop
dependences. I think this makes it more clear that the dependence-type
determines whether task dependences or loop iteration dependences are derived.

Replacing "until specified source iterations" with "until each source
iteration specified by a sink". The original gives the impression that a
single depend clause can specify multiple source iterations, which is not the
case.