make more explicit syntax for source/sink
Before, the syntax for depend clause when used by ordered construct was:

    depend(dependence-type [: vec])

Putting it this way, it wasn't particularly clear that vec is not to be used
for source and it is required for sink. The change here is to treat depend
clause with source and with sink separately, and make it obvious in the syntax
itself whether vec should be used or not.

Under ordered construct section, the clauses may be:

depend(source)
depend(sink : vec)

rather than either of the following which seem to me ambiguous:

depend(dependence-type [: vec])
or
depend(source | sink : vec)