moving descriptions from depend to ordered section
There appeared to be some replication in describing the semantics of ordered
construct with depend clause in the ordered construct section and the depend
clause section. Making this description more clear/precise in the ordered
construct section, and reducing the text describing the sink dependence-type
in the depend clause section.

Also, the note saying vec being a lexicographically later iteration can cause
a deadlock has been amended to say say a deadlock can occur if vec is not a
lexicographically earlier iteration.