Allow impl-defined number of invocations of copy constructors.
In C++, the copy-constructor has to be invoked to initialize a
firstprivate variable. As the implementation can choose to use different
algorithms to do the actually splitting, it may require to copy
construct different numbers of firstprivate variables for the loop tasks
created. This change adds text to be more explicit about this.