cleaning up binding for various device constructs
Removes the binding sections for target and teams. The current defintions for
binding thread set, binding task set, and binding region seem to me
incompatible with these constructs, as they both create new contention groups.

For target update, target enter data, and target exit data, this modifies the
binding task set to to be the generating task, which in this case is actually
the enclosing target task.