clarify data-sharing attributes for dummy args
In Fortran, data-sharing attributes of dummy arguments depends on
whether temporary storage might be used (i.e., "copy-in/copy-out"). This
clarifies that relationship, and specifies conditions under which
temporary storage will be used, will not be used, or its usage is
implementation-defined. This also clarifies conditions under which a data
race may occur when passing a shared variable to a routine due to
potential "copy-out" behavior.