Adding "Miscelaneous Type Definitions" sub-section to chapter 4. This has many, many cross-references, I hope I got them all right.
One big question - is the "ompt_dev_task_none" misnamed? Should it
be instead "ompt_dep_task_none" to indicate that there is no
task dependency? I ask because dev_task_none is used nowhere
in the document, and the task dependency type is missing a "none"
definition like the time, id, data, etc. objects.
If that is the case, it should be renamed and a cross reference
to ompt_task_dependence_t (and vice-versa) should be added.