refining text for dummy argument data-sharing attribute
This refines the description of the dummy argument data-sharing
attribute to capture non-"pass by value" cases where the actual
argument is statically known to be contiguous (hence, no "copy-in"
required).

Also, adds glossary term: simply contiguous array section. This matches
the term found in the Fortran 2008 standard. Since OpenMP still isn't in
sync with F2008, we define it ourselves as an array section that is
statically known to have contiguous storage.