use "write" and "read" flush rather than "release" and "acquire"
This changes the flush properties "release" and "acquire" to "write" and
"read", respectively. A write flush that is synchronizable is said to be a
release flush, and likewise a read flush that is synchronizable is said to be
an acquire flush. This ensures that "release" and "acquire" only pertains to
synchronable operations, in keeping with their meaning in C/C++.