tweak wording after 1st vote
Saying that the corresponding private pointer is initialized with the value of
the base expression in the device data environment is self-referential
(essentially it is saying p = &p[0]).

Change to say that it is initialized "relative to" the address of the
corresponding array section on the device, as opposed to the initial wording
that said it is initialized "with" that address.