remove idle callback
it isn't easy to implement correctly in runtimes that use a half-barrier
at the end of a parallel region.

rather than depending expecting and not getting an idle callback at the
end of a parallel region that uses a half-barrier, it is more reliable
to note ending the of parallel work when entering the terminal barrier.

This commit implements ticket 712