adds directive-variants meta-directive
The directive-variants meta-directive allows directive variants to be
specified and conditionally selected according to the OpenMP context. It uses
the same context selectors as are accepted in the declare variant directive.

As discussed in the May 2018 F2F meeting, the meta-directive takes two forms:
a stand-alone form and a begin/end paired form.

As requested in the 5/29 meeting, the feature was further extended to permit
empty strings to be specified as a directive variant. Additionally, simd
properties may not be specified in a context selector for a directive variant.