Do not restrict which kind of expression can be the base expression of an array section
This restriction in the generic definition of array sections was forbidding
reasonable examples such as:

int* p = ...;
#pragma omp task depend(inout: (p+1)[0:N])
{}

or:

int *f() {}
#pragma omp task depend(inout: f()[0:N])
{}

This restriction was introduced for the map clause. Currently, we have 4 use cases of
array sections:
- Task dependences
- Task affinity
- Reductions
- Map clause

For the first two cases, we do not want to restrict which kind of expression can appear as the base expression
of an array section. For reductions it's already limited to base language identifiers. Finally, the
restriction that was removed from the generic array section has been added to the map clause.