Ticket #776: [OMPT] tracing interface flush runtime entry point
An OpenMP worker thread may finish before a device finalize callback. Before a
worker thread terminates, it needs to be sure that any work that it offloaded
to a device has completed and been reported back to the worker. While a "stop"
command guarantees that device tracing has been flushed, when a worker thread
is finishing, it should not issue a stop command to flush the device because
other threads may still be using it. Thus a flush is needed.