Ticket #775: [OMPT] clarify device finalization
Before the device finalize callback occurs, the device monitoring should be
stopped and all trace data should be flushed back to the tool. The text
associated with the device_finalize_t callback currently says:

A device finalization callback for a tool occurs after all tracing activity on
a device is complete and immediately prior to the finalization of the device
by the OpenMP implementation.