Ticket #765: Fixes in OMPT for flags used as masks
The OMPT interfaces uses flags in forms of masks in several places. To better
distinguish this, the corresponding types had "flag" in the name and the
corresponding constants had hex values. However, this was not consistent and
this ticket proposes to fix this.

This applies to:
- ompt_task_t -> ompt_task_flag_t (and list as hex)
- ompt_dependence_flag_t -> ompt_dependence_type_t (remove hex encoding)
- ompt_mon_flag_t (list as hex)
- ompt_map_flag_t (list as hex)#

Fix parameter names for
- ompt_callback_task_create (CB and record)
- ompt_get_task_info

This ticket also add leading zeros to show proper alignment of masks in the text.