Minor wording change in the cancellation section
A "cancel taskgroup" is not a construct (or at least this is what I think): it is a
cancel construct with a taskgroup clause.

I replaced the second use of "cancel taskgroup" by "that". If it is not clear
enough we can be explicit and write what "that" means in that context.