Removing incomplete list of data-sharing attribute clauses
We were missing 'shared', 'task_reduction' and 'in_reduction' clauses.

Instead of adding them to the list, I rewrote the sentence listing the
cases where it did not apply.

I'd probably revisit this paragraph in a near future, it is not properly written:
we should talk about variables referenced in a list-item that appears in a data sharing clause.