All these restrictions are already present in 2.22.3
There are some repeated restrictions in the firstprivate and lastprivate clauses: we do say that, for those clauses, they inherit
the semantics of the private clause. At the same time, the private clause has the semantics explained in 2.22.3 (List Item privatization).
Thus, we do not really need to list them again.