Moving some content from the Reduction Scoping Clauses to Properties Common To All Reduction Clauses
These sentences describe how private copies are created. In TR4, only Reduction
Scoping Clauses created privated copies. However, at some point we changed
that: nowadays the in_reduction clause, which is a reduction participating
clause, may create private copies.

When we introduced that change we already moved part of the content of the Reduction Scoping clauses section
to the generic one. It seems that we forgot to move some sentences.