RFC: Distinguish between define/declare in omp.h requirements
I assume the intent was that omp.h only declares namespace and and
allocator variables and not that there is one of each per compilation
unit including omp.h.  If so, we should probably say that omp.h
declares it rather than defines it.  On the other hand, even OmpenMP
4.5 says that the file "defines [...] the prototypes of all the
routines..." so maybe wo don't care.  In that case, please disregard
this.

I'm afraid I cannot make any corresponding checks for the Fortran bits
and just have to assume it is correct.