Do not list all categories of required types in omp.h introduction
The introductory paragraph in 3.1 lists a number of categories of
types that omp.h and its Fortran equivalents have to define.  However,
the latest standard version has added a few more so we either have to
grow the list considerably or remove it.  In my opinion, adding the
new type groups to the enumeration would serve no good purpose and
only make the document more difficult to read.  I also did not want to
invent names for the new type categories and therefore decided to
remove it.