misc QC fixes for loop-associated constructs
Remove paragraph that says behavior is unspecified if values used to compute
iteration counts are modified during execution of loops associated with a
loop-associated construct from various sections. This is redundant, since it
is already implied by the constraints on the canonical loop form.

Fix erroneous usage of "parallel loop" in various places.

Fix teams binding description for loop construct, and add missing restrictions
for the loop construct when bind(teams) or bind(parallel) is present.