An attempt to fix constraints on omp_alloc's allocator
Fix wording of restrictions on omp_alloc's allocator argument so that
it reflects the existence of the dynamic_allocators require clause,
makes at least syntactical sense and is consistent with wording
changes introduced by commit 5dd14ff63 elsewhere.  But please see
Jakub's email to the mailing list from 29th August for a list of
questions that are still not resolved.