Do not use math symbols for smaller or greater than in text
I think that using math symbols for greater or smaller than in the
specification text looks rather bad and therefore replaced it with
corresponding text.  Consequently I have also removed dollars so that
the numbers are not typeset with font math.  Please feel free to drop
this patch if uou think it is a bad idea.