ticket 864: mapping pointer-based zero-length array sections
When a pointer-based zero-length array section is mapped, the value of the
privatized pointer is converted from its original value according to a
currently mapped list item to which it is pointing. If a pointer does
not point to an address that is mapped, we just have to check if it
points to an address in the extended range of a currently mapped list
item. If it falls in multiple (overlapped) extended ranges, the behavior
is unspecified -- except for the case where they are all members of the
same structure.

This change also clarifies of use_device_ptr and use_device_addr.